cache compiled SQL statements in gufi_query
In order to not reparse the same SQL statements for each directory,
SQL statements are no longer run during the index walk with
sqlite3_exec. gufi_query now opens a per-thread database in order to
maintain valid SQL prepared statement handles for the duration of the
thread pool. The first time thead[id] runs a SQL statement, it will
pay the cost of parsing to create the handle. All subsequent calls in
thread[id] will reuse the handle. Note that by doing this, each user
provided string will only run the first SQL statement. Any additional
statements will be dropped when preparing the handles.

reoganized gufi_query so that the source file is not so cluttered
    moved code into {include,src}/gufi_query/*

-e has been removed - -K is used to detect aggregation

aggregation causes per-thread files/dbs to not be created

TODO: Convert individual prepared statement handles into lists of handles, and process the entire list.